Added a new iterationsof SidC's notebook.
I have been working on this notebook using the M4 Dataset and manually fitted kernels. However, the models currently have low accuracy, with the best Negative LL only reaching around 2000 after 10000 iterations (I got down to 39 with 40,000 iterations, but I have not verified that recently). I used a combination of RBF, Periodic, and RQ kernels, but other iterations also performed decently. Additionally, I have made modifications to the notebook to graph one of the M4's training and testing time series accurately and provide relevant information. Please refer to the comments in the notebook for more details on all the changes I made. Finally, I have added in some functions that represent relevant tests, but not all of the functions are working.